add scala-js test
