+++ jvm/src/test/scala/scopt/SpecUtilPlatform.scala
@@ -1,0 +1,3 @@
+object SpecUtilPlatform {
+  val envLineSeparator = util.Properties.lineSeparator
+}
