+++ shared/src/test/scala/scopt/SpecUtil.scala
@@ -1,6 +1,6 @@
+import SpecUtilPlatform._
 
 object SpecUtil {
-  val envLineSeparator = util.Properties.lineSeparator
   implicit class RichString(self: String) {
     def newline: String = self + envLineSeparator
     def newlines: String = self.lines.mkString(envLineSeparator)
